Factor out aria.tsx and bulma.ts.
